Refactor components and overall app design and create a state to update final preview of the CV
This commit changed the overall design of the application by removing the following features: (1) The ability to add additional work experiences and educations and (2) the ability to expand/collapse Card components. These changes were made to limit the amount of information the user can input and reduce the overall complexity of the application until further implementations.

This commit also implements the ability to update the final layout of the CV once the user submits their inputs. A useState hook tracks any changes made to each input and a handler function updates the values of the respective elements to display the most recent information the user inputted.

Note: An effort to validate inputs was initially made, but the user can submit empty/invalid inputs as of this commit. A future implementation should focus on validation.